Huge rewrite of tower system + breaking changes (probably) - Every tower is now generated from a config file - Added towers.yml - Rewrite of the Tower.java class - Move some values to doubles since they are superior - GUI are now usable and generated from the config file - More thing that I forgot

+++ src/main/java/me/wyndev/towerdefense/player/IngameTowerDefensePlayer.java
@@ -18,7 +18,7 @@ public class IngameTowerDefensePlayer {
 
     private @Getter TowerDefensePlayer towerDefensePlayer;
     private @Setter int health = 100;
-    private long gold = 8000000; //TODO: set back to 0 for production
+    private double gold = 8000000; //TODO: set back to 0 for production
     private @Setter int towersPlaced = 0;
     /**
      * Gold income of the player every 10 seconds.
@@ -50,7 +50,7 @@ public class IngameTowerDefensePlayer {
         gameSidebar.removeViewer(towerDefensePlayer);
     }
 
-    public void setGold(long gold) {
+    public void setGold(double gold) {
         this.gold = gold;
         this.gameSidebar.updateGoldLine(gold);
     }
